fixing models and controllers

+++ src/main/java/com/example/vol/repositories/PassagerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import javax.transaction.Transactional;
 
 @Transactional()
-public interface PassagerRepository extends JpaRepository<Passager, Long> {
+public interface PassagerRepository extends JpaRepository<Passager, Integer> {
 }
